ccmlib.scylla_node: Simplify scylla startup and shutdown sequence
Right now we use many shell scripts to start scylla and
the JMX service, and the code is ending more complicated
and fragile than we would have hoped.

Get rid of the shell script intermediates and execute
the scylla and scylla-jmx processes directly using
python's subprocess module. Also, get rid of some
sleeps that were put in the code and use better heuristics
to wait for the desired conditions. Adjust the stop procedures
to take the new startup procedure into account.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>